refactor(template-configuration): contract uplift + @ApplicationModule marker
- Lifted 8 contract DTOs to dristi-common/contract/templateconfiguration/
  (OrderPagination, Pagination, TemplateConfiguration, ...Criteria,
  ...ListResponse, ...Request, ...Response, ...SearchRequest)
- Added @ApplicationModule(displayName = "TemplateConfiguration") marker;
  no <Subdomain>Api needed — no DRISTI-internal callers (leaf service)
- No REST conversions (template-configuration is a leaf; _rest_calls.txt empty)
- Cleaned up dead internal.web.models.* wildcard imports left by Phase 35
  alongside the new contract wildcard (4 main + 4 test files)
