+++ dristi-monolith/domain-case-lifecycle/src/main/java/org/pucar/dristi/caselifecycle/templateconfiguration/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * Templateconfiguration subdomain — template configuration storage and retrieval.
+ *
+ * <p>Marked as a Spring Modulith application module. No cross-subdomain
+ * callers exist today; this marker is here for consistency so that when
+ * a caller appears, the boundary is already enforced.
+ */
+@org.springframework.modulith.ApplicationModule(displayName = "TemplateConfiguration")
+package org.pucar.dristi.caselifecycle.templateconfiguration;
