+++ dristi-monolith/dristi-common/src/main/java/org/pucar/dristi/common/contract/templateconfiguration/TemplateConfigurationListResponse.java
@@ -1,4 +1,5 @@
-package org.pucar.dristi.caselifecycle.templateconfiguration.internal.web.models;
+// HAND-CURATED — lifted by Phase 35 (contract-lift)
+package org.pucar.dristi.common.contract.templateconfiguration;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.Valid;
@@ -6,11 +7,15 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
 import org.egov.common.contract.response.ResponseInfo;
 import org.springframework.validation.annotation.Validated;
 
+import java.util.ArrayList;
+import java.util.List;
+
 /**
- * TemplateConfigurationResponse
+ * TemplateConfigurationListResponse
  */
 @Validated
 @jakarta.annotation.Generated(value = "org.egov.codegen.SpringBootCodegen", date = "2024-04-18T11:13:43.389623100+05:30[Asia/Calcutta]")
@@ -18,15 +23,21 @@ import org.springframework.validation.annotation.Validated;
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class TemplateConfigurationResponse {
+public class TemplateConfigurationListResponse {
 
     @JsonProperty("ResponseInfo")
     @Valid
     private ResponseInfo responseInfo = null;
 
-    @JsonProperty("templateConfiguration")
-    @Valid
-    private TemplateConfiguration templateConfiguration = null;
+    @JsonProperty("TotalCount")
+    private Integer totalCount = null;
 
+    @JsonProperty("list")
+    @Valid
+    private List<TemplateConfiguration> list = null;
+
+    @JsonProperty("pagination")
+    @Valid
+    private Pagination pagination = null;
 
 }
